+++ R/evinb.R
@@ -382,7 +382,7 @@ evinb <- function(
   verbose = FALSE
 ) {
   i <- 'temp_iter'
-  block <- evinf_block_name(rlang::enquo(block), parent.frame())
+  block <- evinf_block_name(rlang::enquo(block), parent.frame(), data)
   mc <- match.call()
   ctrl <- resolve_evinf_control(control, mc, environment(), fn = "evinb")